refactor: split large files and reorganize ai/ & db/ directory structure
- Split tools.ts (~1748 lines) into 6 domain modules:
  rag-tools, analysis-tools, annotation-tools,
  library-tools, skill-tools, mindmap-tools

- Move tool modules into ai/tools/ subdirectory with barrel index

- Move test files into __tests__/ subdirectories (ai & db)

- Split database.ts into domain-specific query modules:
  book-queries, chunk-queries, highlight-queries, note-queries,
  bookmark-queries, thread-queries, message-queries,
  session-queries, skill-queries, db-core

- Update all import paths across packages (62 references)

- All 117 tests passing (10 test files)
