+++ packages/core/src/ai/tools/tool-types.ts
