+++ packages/core/src/db/database.ts
@@ -1,1893 +1,103 @@
-import type { IDatabase } from "../services/platform";
-import { getPlatformService } from "../services/platform";
-import { runSerializedDbTask } from "./write-retry";
 /**
- * Database access layer — platform-agnostic via IDatabase interface
- * Uses getPlatformService().loadDatabase() to obtain a database connection.
- * No direct dependency on Tauri or any platform-specific package.
+ * Database access layer — re-exports from domain-specific modules.
+ *
+ * This file exists for backward compatibility: external modules that
+ * `import { … } from "./database"` will continue to work unchanged.
+ *
+ * New code should import from the specific query module or from the
+ * barrel `./index` (equivalently `../db`).
  */
-import type { Book, Bookmark, Chunk, Highlight, Message, Note, Skill, Thread } from "../types";
-import type { ReadingSession } from "../types/reading";
-import { generateId } from "../utils/generate-id";
 
-// Lazy-loaded database instances
-let db: IDatabase | null = null;
-let dbInitialized = false;
-let dbInitPromise: Promise<void> | null = null;
-let dbLoadPromise: Promise<IDatabase> | null = null;
-
-let localDb: IDatabase | null = null;
-let localDbInitialized = false;
-let localDbInitPromise: Promise<void> | null = null;
-let localDbLoadPromise: Promise<IDatabase> | null = null;
-
-const DB_FILENAME = "readany.db";
-const LOCAL_DB_FILENAME = "readany_local.db";
-const DEVICE_ID_STORAGE_KEY = "sync_device_id";
-const DESKTOP_DATA_ROOT_CONFIG_FILE = "desktop-data-root.json";
-
-// Cached device ID for sync tracking
-let cachedDeviceId: string | null = null;
-
-function normalizeDir(path: string): string {
-  const trimmed = path.replace(/^file:\/\//, "").trim();
-  if (!trimmed) return "";
-  if (/^[A-Za-z]:\\$/.test(trimmed)) return trimmed;
-  return trimmed.replace(/[\\/]+$/, "");
-}
-
-async function getDefaultDataRoot(): Promise<string> {
-  const platform = getPlatformService();
-  return normalizeDir(await platform.getAppDataDir());
-}
-
-async function getDesktopDataRootConfigPath(): Promise<string> {
-  const platform = getPlatformService();
-  return platform.joinPath(await getDefaultDataRoot(), DESKTOP_DATA_ROOT_CONFIG_FILE);
-}
-
-async function getDesktopDataRoot(): Promise<string> {
-  const platform = getPlatformService();
-  const defaultRoot = await getDefaultDataRoot();
-  if (!platform.isDesktop) {
-    return defaultRoot;
-  }
-
-  try {
-    const configPath = await getDesktopDataRootConfigPath();
-    if (!(await platform.exists(configPath))) {
-      return defaultRoot;
-    }
-
-    const raw = await platform.readTextFile(configPath);
-    const parsed = JSON.parse(raw) as { dataRoot?: string };
-    const configuredRoot = normalizeDir(parsed.dataRoot || "");
-    return configuredRoot || defaultRoot;
-  } catch {
-    return defaultRoot;
-  }
-}
-
-export async function getActiveDataRoot(): Promise<string> {
-  return getDesktopDataRoot();
-}
-
-export async function getDatabaseFilePath(filename: string): Promise<string> {
-  const platform = getPlatformService();
-  return platform.joinPath(await getDesktopDataRoot(), filename);
-}
-
-async function getDatabaseLocation(filename: string): Promise<string> {
-  const platform = getPlatformService();
-  if (!platform.isDesktop) {
-    return filename;
-  }
-
-  return `sqlite:${await getDatabaseFilePath(filename)}`;
-}
-
-async function configureDatabaseConnection(database: IDatabase): Promise<void> {
-  try {
-    await database.execute("PRAGMA journal_mode = WAL");
-  } catch {
-    // Some adapters may not support WAL mode changes
-  }
-  try {
-    await database.execute("PRAGMA synchronous = NORMAL");
-  } catch {
-    // Some adapters may not support synchronous mode changes
-  }
-  try {
-    await database.execute("PRAGMA foreign_keys = ON");
-  } catch {
-    // Some adapters may not support PRAGMA configuration
-  }
-  try {
-    await database.execute("PRAGMA busy_timeout = 15000");
-  } catch {
-    // Some adapters may not support PRAGMA configuration
-  }
-}
-
-export async function cleanupOrphanedSyncRows(databaseArg?: IDatabase): Promise<void> {
-  const database = databaseArg ?? (await getDB());
-
-  const cleanupStatements = [
-    "DELETE FROM highlights WHERE book_id NOT IN (SELECT id FROM books)",
-    "DELETE FROM notes WHERE book_id NOT IN (SELECT id FROM books)",
-    "DELETE FROM bookmarks WHERE book_id NOT IN (SELECT id FROM books)",
-    "DELETE FROM reading_sessions WHERE book_id NOT IN (SELECT id FROM books)",
-    "DELETE FROM book_tags WHERE book_id NOT IN (SELECT id FROM books) OR tag_id NOT IN (SELECT id FROM tags)",
-    "DELETE FROM messages WHERE thread_id NOT IN (SELECT id FROM threads)",
-  ];
-
-  for (const sql of cleanupStatements) {
-    try {
-      await database.execute(sql);
-    } catch {
-      // Ignore partial cleanup failures on older schema variants.
-    }
-  }
-}
-
-export async function getDB(): Promise<IDatabase> {
-  if (db) return db;
-
-  if (!dbLoadPromise) {
-    dbLoadPromise = (async () => {
-      const platform = getPlatformService();
-      const loadedDb = await platform.loadDatabase(await getDatabaseLocation(DB_FILENAME));
-      await configureDatabaseConnection(loadedDb);
-      db = loadedDb;
-      return loadedDb;
-    })().finally(() => {
-      dbLoadPromise = null;
-    });
-  }
-
-  return dbLoadPromise;
-}
-
-/** Get or lazily open the local database (readany_local.db) */
-export async function getLocalDB(): Promise<IDatabase> {
-  if (localDb) return localDb;
-
-  if (!localDbLoadPromise) {
-    localDbLoadPromise = (async () => {
-      const platform = getPlatformService();
-      const loadedDb = await platform.loadDatabase(await getDatabaseLocation(LOCAL_DB_FILENAME));
-      await configureDatabaseConnection(loadedDb);
-      localDb = loadedDb;
-      return loadedDb;
-    })().finally(() => {
-      localDbLoadPromise = null;
-    });
-  }
-
-  return localDbLoadPromise;
-}
-
-/** Close the active database connection and clear cache */
-export async function closeDB(): Promise<void> {
-  dbInitPromise = null;
-  dbLoadPromise = null;
-  if (db) {
-    try {
-      await db.close();
-    } catch {
-      // Ignore close errors
-    }
-    db = null;
-    dbInitialized = false;
-    cachedDeviceId = null;
-  }
-  // Also close the local database
-  await closeLocalDB();
-}
-
-/** Close the local database connection and clear cache */
-export async function closeLocalDB(): Promise<void> {
-  localDbInitPromise = null;
-  localDbLoadPromise = null;
-  if (localDb) {
-    try {
-      await localDb.close();
-    } catch {
-      // Ignore close errors
-    }
-    localDb = null;
-    localDbInitialized = false;
-  }
-}
-
-/** Ensure no active transaction (rollback if any) */
-export async function ensureNoTransaction(): Promise<void> {
-  if (db) {
-    try {
-      await db.execute("ROLLBACK", []);
-    } catch {
-      // No active transaction, ignore
-    }
-  }
-}
-
-/** Reset the DB cache without closing (for use after external file replacement) */
-export function resetDBCache(): void {
-  db = null;
-  dbInitialized = false;
-  cachedDeviceId = null;
-}
-
-/** Reset the local DB cache without closing */
-export function resetLocalDBCache(): void {
-  localDb = null;
-  localDbInitialized = false;
-}
-
-/** Get or create device ID for sync tracking */
-export async function getDeviceId(): Promise<string> {
-  if (cachedDeviceId) return cachedDeviceId;
-
-  const platform = getPlatformService();
-  let kvAvailable = true;
-
-  try {
-    const storedDeviceId = await platform.kvGetItem(DEVICE_ID_STORAGE_KEY);
-    if (storedDeviceId) {
-      cachedDeviceId = storedDeviceId;
-      try {
-        const database = await getDB();
-        await database.execute(
-          "INSERT OR REPLACE INTO sync_metadata (key, value) VALUES ('device_id', ?)",
-          [storedDeviceId],
-        );
-      } catch {
-        // Table might not exist yet during init
-      }
-      return storedDeviceId;
-    }
-  } catch {
-    kvAvailable = false;
-  }
-
-  const database = await getDB();
-  if (!kvAvailable) {
-    try {
-      const rows = await database.select<{ value: string }>(
-        "SELECT value FROM sync_metadata WHERE key = 'device_id'",
-      );
-      if (rows.length > 0 && rows[0].value) {
-        cachedDeviceId = rows[0].value;
-        return rows[0].value;
-      }
-    } catch {
-      // Table might not exist yet
-    }
-  }
-
-  // Generate new device ID
-  const id = generateId();
-  try {
-    await database.execute(
-      "INSERT OR REPLACE INTO sync_metadata (key, value) VALUES ('device_id', ?)",
-      [id],
-    );
-  } catch {
-    // Table might not exist yet during init
-  }
-  try {
-    await platform.kvSetItem(DEVICE_ID_STORAGE_KEY, id);
-  } catch {
-    // Ignore KV persistence errors; DB copy will still exist.
-  }
-  cachedDeviceId = id;
-  return id;
-}
-
-/** Get next sync version for a table */
-async function nextSyncVersion(database: IDatabase, table: string): Promise<number> {
-  const rows = await database.select<{ max_v: number | null }>(
-    `SELECT MAX(sync_version) as max_v FROM ${table}`,
-  );
-  return (rows[0]?.max_v || 0) + 1;
-}
-
-async function nextUpdatedAt(database: IDatabase, table: string, id: string): Promise<number> {
-  const now = Date.now();
-
-  try {
-    const rows = await database.select<{ updated_at: number | null }>(
-      `SELECT updated_at FROM ${table} WHERE id = ?`,
-      [id],
-    );
-    const current = rows[0]?.updated_at ?? 0;
-    return Math.max(now, current + 1);
-  } catch {
-    return now;
-  }
-}
-
-/** Insert a tombstone record for sync deletion tracking */
-async function insertTombstone(database: IDatabase, id: string, tableName: string): Promise<void> {
-  const deviceId = await getDeviceId();
-  try {
-    await database.execute(
-      "INSERT OR REPLACE INTO sync_tombstones (id, table_name, deleted_at, device_id) VALUES (?, ?, ?, ?)",
-      [id, tableName, Date.now(), deviceId],
-    );
-  } catch {
-    // sync_tombstones table might not exist on older schema
-  }
-}
-
-/** Initialize the database, creating tables if needed */
-export async function initDatabase(): Promise<void> {
-  if (dbInitialized) return;
-  if (dbInitPromise) return dbInitPromise;
-
-  dbInitPromise = (async () => {
-    await runSerializedDbTask(async () => {
-      if (dbInitialized) return;
-
-      const database = await getDB();
-
-      // Create tables
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS books (
-      id TEXT PRIMARY KEY,
-      file_path TEXT NOT NULL,
-      format TEXT NOT NULL DEFAULT 'epub',
-      title TEXT NOT NULL DEFAULT '',
-      author TEXT NOT NULL DEFAULT '',
-      publisher TEXT,
-      language TEXT,
-      isbn TEXT,
-      description TEXT,
-      cover_url TEXT,
-      publish_date TEXT,
-      subjects TEXT,
-      total_pages INTEGER DEFAULT 0,
-      total_chapters INTEGER DEFAULT 0,
-      added_at INTEGER NOT NULL,
-      last_opened_at INTEGER,
-      progress REAL DEFAULT 0,
-      current_cfi TEXT,
-      is_vectorized INTEGER DEFAULT 0,
-      vectorize_progress REAL DEFAULT 0,
-      tags TEXT DEFAULT '[]'
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS highlights (
-      id TEXT PRIMARY KEY,
-      book_id TEXT NOT NULL,
-      cfi TEXT NOT NULL,
-      text TEXT NOT NULL,
-      color TEXT NOT NULL DEFAULT 'yellow',
-      note TEXT,
-      chapter_title TEXT,
-      created_at INTEGER NOT NULL,
-      updated_at INTEGER NOT NULL,
-      FOREIGN KEY (book_id) REFERENCES books(id) ON DELETE CASCADE
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS notes (
-      id TEXT PRIMARY KEY,
-      book_id TEXT NOT NULL,
-      highlight_id TEXT,
-      cfi TEXT,
-      title TEXT NOT NULL DEFAULT '',
-      content TEXT NOT NULL DEFAULT '',
-      chapter_title TEXT,
-      tags TEXT DEFAULT '[]',
-      created_at INTEGER NOT NULL,
-      updated_at INTEGER NOT NULL,
-      FOREIGN KEY (book_id) REFERENCES books(id) ON DELETE CASCADE
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS bookmarks (
-      id TEXT PRIMARY KEY,
-      book_id TEXT NOT NULL,
-      cfi TEXT NOT NULL,
-      label TEXT,
-      chapter_title TEXT,
-      created_at INTEGER NOT NULL,
-      FOREIGN KEY (book_id) REFERENCES books(id) ON DELETE CASCADE
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS tags (
-      id TEXT PRIMARY KEY,
-      name TEXT NOT NULL UNIQUE,
-      updated_at INTEGER NOT NULL DEFAULT 0
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS book_tags (
-      id TEXT PRIMARY KEY,
-      book_id TEXT NOT NULL,
-      tag_id TEXT NOT NULL,
-      updated_at INTEGER NOT NULL DEFAULT 0,
-      UNIQUE (book_id, tag_id),
-      FOREIGN KEY (book_id) REFERENCES books(id) ON DELETE CASCADE,
-      FOREIGN KEY (tag_id) REFERENCES tags(id) ON DELETE CASCADE
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS threads (
-      id TEXT PRIMARY KEY,
-      book_id TEXT,
-      title TEXT NOT NULL DEFAULT '',
-      created_at INTEGER NOT NULL,
-      updated_at INTEGER NOT NULL
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS messages (
-      id TEXT PRIMARY KEY,
-      thread_id TEXT NOT NULL,
-      role TEXT NOT NULL,
-      content TEXT NOT NULL DEFAULT '',
-      citations TEXT,
-      tool_calls TEXT,
-      reasoning TEXT,
-      created_at INTEGER NOT NULL,
-      FOREIGN KEY (thread_id) REFERENCES threads(id) ON DELETE CASCADE
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS skills (
-      id TEXT PRIMARY KEY,
-      name TEXT NOT NULL,
-      description TEXT NOT NULL DEFAULT '',
-      icon TEXT,
-      enabled INTEGER DEFAULT 1,
-      parameters TEXT DEFAULT '[]',
-      prompt TEXT DEFAULT '',
-      built_in INTEGER DEFAULT 0,
-      created_at INTEGER NOT NULL,
-      updated_at INTEGER NOT NULL
-    )
-  `);
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS reading_sessions (
-      id TEXT PRIMARY KEY,
-      book_id TEXT NOT NULL,
-      started_at INTEGER NOT NULL,
-      ended_at INTEGER,
-      total_active_time INTEGER DEFAULT 0,
-      pages_read INTEGER DEFAULT 0,
-      state TEXT DEFAULT 'active',
-      updated_at INTEGER NOT NULL DEFAULT 0,
-      FOREIGN KEY (book_id) REFERENCES books(id) ON DELETE CASCADE
-    )
-  `);
-
-      // Create indexes
-      await database.execute("CREATE INDEX IF NOT EXISTS idx_highlights_book ON highlights(book_id)");
-      await database.execute("CREATE INDEX IF NOT EXISTS idx_notes_book ON notes(book_id)");
-      await database.execute("CREATE INDEX IF NOT EXISTS idx_bookmarks_book ON bookmarks(book_id)");
-      await database.execute("CREATE INDEX IF NOT EXISTS idx_messages_thread ON messages(thread_id)");
-      await database.execute(
-        "CREATE INDEX IF NOT EXISTS idx_reading_sessions_book ON reading_sessions(book_id)",
-      );
-
-      // Migrations: add columns that may be missing from older schema versions
-      try {
-        await database.execute("ALTER TABLE books ADD COLUMN format TEXT NOT NULL DEFAULT 'epub'");
-      } catch {
-        // Column already exists, ignore
-      }
-      try {
-        await database.execute("ALTER TABLE books ADD COLUMN tags TEXT DEFAULT '[]'");
-      } catch {
-        // Column already exists, ignore
-      }
-      try {
-        await database.execute("ALTER TABLE messages ADD COLUMN reasoning TEXT");
-      } catch {
-        // Column already exists, ignore
-      }
-      try {
-        await database.execute("ALTER TABLE messages ADD COLUMN parts_order TEXT");
-      } catch {
-        // Column already exists, ignore
-      }
-      // --- Sync migrations ---
-      // Migration 4: Add updated_at and file_hash to books
-      try {
-        await database.execute("ALTER TABLE books ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0");
-      } catch {
-        // Column already exists
-      }
-      try {
-        await database.execute("ALTER TABLE books ADD COLUMN file_hash TEXT");
-      } catch {
-        // Column already exists
-      }
-      try {
-        await database.execute("UPDATE books SET updated_at = added_at WHERE updated_at = 0");
-      } catch {
-        // Already updated
-      }
-
-      // Migration 5: Tombstones table
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS sync_tombstones (
-      id TEXT NOT NULL,
-      table_name TEXT NOT NULL,
-      deleted_at INTEGER NOT NULL,
-      device_id TEXT NOT NULL,
-      PRIMARY KEY (id, table_name)
-    )
-  `);
-      await database.execute(
-        "CREATE INDEX IF NOT EXISTS idx_tombstones_deleted_at ON sync_tombstones(deleted_at)",
-      );
-
-      // Migration 6: Sync metadata table
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS sync_metadata (
-      key TEXT PRIMARY KEY,
-      value TEXT NOT NULL
-    )
-  `);
-
-      // Migration 7: Add sync_version and last_modified_by to all synced tables
-      const syncTables = [
-        "books",
-        "highlights",
-        "notes",
-        "bookmarks",
-        "tags",
-        "book_tags",
-        "reading_sessions",
-        "threads",
-        "messages",
-        "skills",
-      ];
-      for (const table of syncTables) {
-        try {
-          await database.execute(`ALTER TABLE ${table} ADD COLUMN sync_version INTEGER DEFAULT 0`);
-        } catch {
-          // Column already exists
-        }
-        try {
-          await database.execute(`ALTER TABLE ${table} ADD COLUMN last_modified_by TEXT`);
-        } catch {
-          // Column already exists
-        }
-      }
-
-      // Migration 8: Add updated_at to tables that need it for incremental sync
-      const tablesNeedingUpdatedAt = ["bookmarks"];
-      for (const table of tablesNeedingUpdatedAt) {
-        try {
-          await database.execute(
-            `ALTER TABLE ${table} ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0`,
-          );
-        } catch {
-          // Column already exists
-        }
-      }
-      // Initialize updated_at from created_at
-      try {
-        await database.execute("UPDATE bookmarks SET updated_at = created_at WHERE updated_at = 0");
-      } catch {
-        // Already updated or column doesn't exist
-      }
-
-      // Migration 9: Add sync_status to books for on-demand download
-      try {
-        await database.execute(
-          "ALTER TABLE books ADD COLUMN sync_status TEXT NOT NULL DEFAULT 'local'",
-        );
-      } catch {
-        // Column already exists
-      }
-
-      // Migration 10: Add updated_at and id to tags/book_tags for sync; add updated_at to reading_sessions
-      try {
-        await database.execute("ALTER TABLE tags ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0");
-      } catch {
-        // Column already exists
-      }
-      try {
-        await database.execute("ALTER TABLE book_tags ADD COLUMN id TEXT");
-        // Backfill id for existing rows (book_id || '-' || tag_id as stable key)
-        await database.execute(
-          "UPDATE book_tags SET id = book_id || '-' || tag_id WHERE id IS NULL",
-        );
-      } catch {
-        // Column already exists
-      }
-      try {
-        await database.execute(
-          "ALTER TABLE book_tags ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0",
-        );
-      } catch {
-        // Column already exists
-      }
-      try {
-        await database.execute(
-          "ALTER TABLE reading_sessions ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0",
-        );
-        await database.execute(
-          "UPDATE reading_sessions SET updated_at = started_at WHERE updated_at = 0",
-        );
-      } catch {
-        // Column already exists or table doesn't exist yet
-      }
-
-      const platform = getPlatformService();
-      if (platform.isDesktop) {
-        await cleanupOrphanedSyncRows(database);
-      }
-
-      dbInitialized = true;
-    });
-
-    // Also initialize the local database (chunks only) after the main DB init
-    // finishes, so schema migration and sync apply never compete for writes.
-    await initLocalDatabase();
-  })().finally(() => {
-    dbInitPromise = null;
-  });
-
-  return dbInitPromise;
-}
-export async function initLocalDatabase(): Promise<void> {
-  if (localDbInitialized) return;
-  if (localDbInitPromise) return localDbInitPromise;
-
-  localDbInitPromise = (async () => {
-    await runSerializedDbTask(async () => {
-      if (localDbInitialized) return;
-
-      const database = await getLocalDB();
-
-      await database.execute(`
-    CREATE TABLE IF NOT EXISTS chunks (
-      id TEXT PRIMARY KEY,
-      book_id TEXT NOT NULL,
-      chapter_index INTEGER NOT NULL,
-      chapter_title TEXT NOT NULL DEFAULT '',
-      content TEXT NOT NULL,
-      token_count INTEGER NOT NULL DEFAULT 0,
-      start_cfi TEXT,
-      end_cfi TEXT,
-      segment_cfis TEXT,
-      embedding BLOB,
-      updated_at INTEGER NOT NULL DEFAULT 0
-    )
-  `);
-
-      try {
-        await database.execute(
-          "ALTER TABLE chunks ADD COLUMN updated_at INTEGER NOT NULL DEFAULT 0",
-        );
-      } catch {
-        // Column already exists on upgraded installs.
-      }
-
-      // Create indexes
-      await database.execute("CREATE INDEX IF NOT EXISTS idx_chunks_book ON chunks(book_id)");
-
-      localDbInitialized = true;
-
-      // Migrate data from main DB to local DB on first run
-      await migrateDataToLocalDB();
-    });
-  })().finally(() => {
-    localDbInitPromise = null;
-  });
-
-  return localDbInitPromise;
-}
-
-/** Migrate chunks from main DB to local DB (one-time) */
-async function migrateDataToLocalDB(): Promise<void> {
-  const platform = getPlatformService();
-  if (!platform.isDesktop) {
-    return;
-  }
-
-  const mainDB = await getDB();
-
-  // Check if migration has already been done
-  try {
-    const rows = await mainDB.select<{ value: string }>(
-      "SELECT value FROM sync_metadata WHERE key = 'local_db_migration_done'",
-    );
-    if (rows.length > 0 && rows[0].value === "1") {
-      return; // Already migrated
-    }
-  } catch {
-    // sync_metadata might not exist yet, skip migration
-    return;
-  }
-
-  // Check if chunks table exists in main DB
-  let hasChunksInMain = false;
-
-  try {
-    await mainDB.select<{ id: string }>("SELECT id FROM chunks LIMIT 1");
-    hasChunksInMain = true;
-  } catch {
-    // Table doesn't exist in main DB
-  }
-
-  const localDB = await getLocalDB();
-
-  // Migrate chunks
-  if (hasChunksInMain) {
-    try {
-      const chunks = await mainDB.select<Record<string, unknown>>(
-        "SELECT id, book_id, chapter_index, chapter_title, content, token_count, start_cfi, end_cfi, segment_cfis, embedding FROM chunks",
-      );
-      for (const chunk of chunks) {
-        await localDB.execute(
-          "INSERT OR IGNORE INTO chunks (id, book_id, chapter_index, chapter_title, content, token_count, start_cfi, end_cfi, segment_cfis, embedding) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-          [
-            chunk.id,
-            chunk.book_id,
-            chunk.chapter_index,
-            chunk.chapter_title,
-            chunk.content,
-            chunk.token_count,
-            chunk.start_cfi,
-            chunk.end_cfi,
-            chunk.segment_cfis,
-            chunk.embedding,
-          ],
-        );
-      }
-      // Drop from main DB
-      await mainDB.execute("DROP TABLE IF EXISTS chunks");
-    } catch {
-      // Migration error — non-fatal, table may be partially migrated
-    }
-  }
-
-  // Mark migration as done
-  try {
-    await mainDB.execute(
-      "INSERT OR REPLACE INTO sync_metadata (key, value) VALUES ('local_db_migration_done', '1')",
-    );
-  } catch {
-    // Non-fatal
-  }
-}
-
-function serializeEmbedding(embedding?: number[]): Uint8Array | null {
-  if (!embedding || embedding.length === 0) return null;
-  const buffer = new ArrayBuffer(embedding.length * 4);
-  const view = new Float32Array(buffer);
-  for (let i = 0; i < embedding.length; i++) {
-    view[i] = embedding[i];
-  }
-  return new Uint8Array(buffer);
-}
-
-function deserializeEmbedding(data: unknown): number[] | undefined {
-  if (!data) return undefined;
-  // Data comes as an array of bytes from the SQL plugin
-  const bytes = data instanceof Uint8Array ? data : new Uint8Array(data as ArrayBuffer);
-  if (bytes.length === 0) return undefined;
-  const view = new Float32Array(bytes.buffer, bytes.byteOffset, bytes.byteLength / 4);
-  return Array.from(view);
-}
-
-function parseJSON<T>(str: string | null | undefined, fallback: T): T {
-  if (!str) return fallback;
-  try {
-    return JSON.parse(str) as T;
-  } catch {
-    return fallback;
-  }
-}
-
-// --- Books ---
-
-interface BookRow {
-  id: string;
-  file_path: string;
-  format: string;
-  title: string;
-  author: string;
-  publisher: string | null;
-  language: string | null;
-  isbn: string | null;
-  description: string | null;
-  cover_url: string | null;
-  publish_date: string | null;
-  subjects: string | null;
-  total_pages: number;
-  total_chapters: number;
-  added_at: number;
-  last_opened_at: number | null;
-  updated_at: number;
-  progress: number;
-  current_cfi: string | null;
-  is_vectorized: number;
-  vectorize_progress: number;
-  tags: string;
-  file_hash: string | null;
-  sync_status: string;
-}
-
-function rowToBook(row: BookRow): Book {
-  return {
-    id: row.id,
-    filePath: row.file_path,
-    format: (row.format as Book["format"]) || "epub",
-    meta: {
-      title: row.title,
-      author: row.author,
-      publisher: row.publisher || undefined,
-      language: row.language || undefined,
-      isbn: row.isbn || undefined,
-      description: row.description || undefined,
-      coverUrl: row.cover_url || undefined,
-      publishDate: row.publish_date || undefined,
-      subjects: parseJSON(row.subjects, undefined),
-      totalPages: row.total_pages || undefined,
-      totalChapters: row.total_chapters || undefined,
-    },
-    addedAt: row.added_at,
-    lastOpenedAt: row.last_opened_at || undefined,
-    updatedAt: row.updated_at || row.added_at,
-    progress: row.progress,
-    currentCfi: row.current_cfi || undefined,
-    isVectorized: row.is_vectorized === 1,
-    vectorizeProgress: row.vectorize_progress,
-    tags: parseJSON(row.tags, []),
-    fileHash: row.file_hash || undefined,
-    syncStatus: (row.sync_status as Book["syncStatus"]) || "local",
-  };
-}
-
-export async function getBooks(): Promise<Book[]> {
-  const database = await getDB();
-  const rows = await database.select<BookRow>(
-    "SELECT * FROM books ORDER BY last_opened_at DESC, added_at DESC",
-  );
-  return rows.map(rowToBook);
-}
-
-export async function getBook(id: string): Promise<Book | null> {
-  const database = await getDB();
-  const rows = await database.select<BookRow>("SELECT * FROM books WHERE id = ?", [id]);
-  return rows.length > 0 ? rowToBook(rows[0]) : null;
-}
-
-export async function insertBook(book: Book): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "books");
-  const now = Date.now();
-  await database.execute(
-    `INSERT INTO books (id, file_path, format, title, author, publisher, language, isbn, description, cover_url, publish_date, subjects, total_pages, total_chapters, added_at, last_opened_at, updated_at, progress, current_cfi, is_vectorized, vectorize_progress, tags, file_hash, sync_status, sync_version, last_modified_by)
-     VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-    [
-      book.id,
-      book.filePath,
-      book.format || "epub",
-      book.meta.title,
-      book.meta.author,
-      book.meta.publisher || null,
-      book.meta.language || null,
-      book.meta.isbn || null,
-      book.meta.description || null,
-      book.meta.coverUrl || null,
-      book.meta.publishDate || null,
-      book.meta.subjects ? JSON.stringify(book.meta.subjects) : null,
-      book.meta.totalPages || 0,
-      book.meta.totalChapters || 0,
-      book.addedAt,
-      book.lastOpenedAt || null,
-      now,
-      book.progress,
-      book.currentCfi || null,
-      book.isVectorized ? 1 : 0,
-      book.vectorizeProgress,
-      JSON.stringify(book.tags),
-      book.fileHash || null,
-      book.syncStatus || "local",
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function updateBook(id: string, updates: Partial<Book>): Promise<void> {
-  const database = await getDB();
-  const sets: string[] = [];
-  const values: unknown[] = [];
-
-  if (updates.filePath !== undefined) {
-    sets.push("file_path = ?");
-    values.push(updates.filePath);
-  }
-  if (updates.meta?.title !== undefined) {
-    sets.push("title = ?");
-    values.push(updates.meta.title);
-  }
-  if (updates.meta?.author !== undefined) {
-    sets.push("author = ?");
-    values.push(updates.meta.author);
-  }
-  if (updates.meta?.coverUrl !== undefined) {
-    sets.push("cover_url = ?");
-    values.push(updates.meta.coverUrl || null);
-  }
-  if (updates.meta?.publisher !== undefined) {
-    sets.push("publisher = ?");
-    values.push(updates.meta.publisher || null);
-  }
-  if (updates.meta?.description !== undefined) {
-    sets.push("description = ?");
-    values.push(updates.meta.description || null);
-  }
-  if (updates.meta?.language !== undefined) {
-    sets.push("language = ?");
-    values.push(updates.meta.language || null);
-  }
-  if (updates.meta?.totalPages !== undefined) {
-    sets.push("total_pages = ?");
-    values.push(updates.meta.totalPages);
-  }
-  if (updates.format !== undefined) {
-    sets.push("format = ?");
-    values.push(updates.format);
-  }
-  if (updates.progress !== undefined) {
-    sets.push("progress = ?");
-    values.push(updates.progress);
-  }
-  if (updates.currentCfi !== undefined) {
-    sets.push("current_cfi = ?");
-    values.push(updates.currentCfi);
-  }
-  if (updates.lastOpenedAt !== undefined) {
-    sets.push("last_opened_at = ?");
-    values.push(updates.lastOpenedAt);
-  }
-  if (updates.isVectorized !== undefined) {
-    sets.push("is_vectorized = ?");
-    values.push(updates.isVectorized ? 1 : 0);
-  }
-  if (updates.vectorizeProgress !== undefined) {
-    sets.push("vectorize_progress = ?");
-    values.push(updates.vectorizeProgress);
-  }
-  if (updates.tags !== undefined) {
-    sets.push("tags = ?");
-    values.push(JSON.stringify(updates.tags));
-  }
-  if (updates.fileHash !== undefined) {
-    sets.push("file_hash = ?");
-    values.push(updates.fileHash);
-  }
-  if (updates.syncStatus !== undefined) {
-    sets.push("sync_status = ?");
-    values.push(updates.syncStatus);
-  }
-
-  if (sets.length === 0) return;
-
-  // Add sync tracking fields
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "books");
-  const updatedAt = await nextUpdatedAt(database, "books", id);
-  sets.push("updated_at = ?");
-  values.push(updatedAt);
-  sets.push("sync_version = ?");
-  values.push(syncVersion);
-  sets.push("last_modified_by = ?");
-  values.push(deviceId);
-
-  values.push(id);
-  await database.execute(`UPDATE books SET ${sets.join(", ")} WHERE id = ?`, values);
-}
-
-export async function setBookSyncStatus(id: string, syncStatus: Book["syncStatus"]): Promise<void> {
-  const database = await getDB();
-  await database.execute("UPDATE books SET sync_status = ? WHERE id = ?", [syncStatus, id]);
-}
-
-export async function deleteBook(id: string): Promise<void> {
-  const database = await getDB();
-
-  // Delete all related data first to maintain referential integrity
-  // Delete highlights
-  await database.execute("DELETE FROM highlights WHERE book_id = ?", [id]);
-
-  // Delete bookmarks
-  await database.execute("DELETE FROM bookmarks WHERE book_id = ?", [id]);
-
-  // Delete reading sessions
-  await database.execute("DELETE FROM reading_sessions WHERE book_id = ?", [id]);
-
-  // Delete chunks from local DB
-  const localDatabase = await getLocalDB();
-  await localDatabase.execute("DELETE FROM chunks WHERE book_id = ?", [id]);
-
-  // Finally, delete the book itself
-  await insertTombstone(database, id, "books");
-  await database.execute("DELETE FROM books WHERE id = ?", [id]);
-}
-
-// --- Highlights ---
-
-interface HighlightRow {
-  id: string;
-  book_id: string;
-  cfi: string;
-  text: string;
-  color: string;
-  note: string | null;
-  chapter_title: string | null;
-  created_at: number;
-  updated_at: number;
-}
-
-export async function getHighlights(bookId: string): Promise<Highlight[]> {
-  const database = await getDB();
-  const rows = await database.select<HighlightRow>(
-    "SELECT * FROM highlights WHERE book_id = ? ORDER BY created_at DESC",
-    [bookId],
-  );
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    cfi: r.cfi,
-    text: r.text,
-    color: r.color as Highlight["color"],
-    note: r.note || undefined,
-    chapterTitle: r.chapter_title || undefined,
-    createdAt: r.created_at,
-    updatedAt: r.updated_at,
-  }));
-}
-
-/** Get all highlights across all books (for general chat without bookId) */
-export async function getAllHighlights(limit = 50): Promise<Highlight[]> {
-  const database = await getDB();
-  const rows = await database.select<HighlightRow>(
-    "SELECT * FROM highlights ORDER BY created_at DESC LIMIT ?",
-    [limit],
-  );
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    cfi: r.cfi,
-    text: r.text,
-    color: r.color as Highlight["color"],
-    note: r.note || undefined,
-    chapterTitle: r.chapter_title || undefined,
-    createdAt: r.created_at,
-    updatedAt: r.updated_at,
-  }));
-}
-
-/** Extended highlight with book info for notes page */
-export interface HighlightWithBook extends Highlight {
-  bookTitle: string;
-  bookAuthor: string;
-  bookCoverUrl?: string;
-}
-
-/** Get all highlights with book info (JOIN query) */
-export async function getAllHighlightsWithBooks(limit = 500): Promise<HighlightWithBook[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    cfi: string;
-    text: string;
-    color: string;
-    note: string | null;
-    chapter_title: string | null;
-    created_at: number;
-    updated_at: number;
-    book_title: string;
-    book_author: string;
-    book_cover_url: string | null;
-  }>(
-    `SELECT 
-      h.id, h.book_id, h.cfi, h.text, h.color, h.note, h.chapter_title, h.created_at, h.updated_at,
-      b.title as book_title, b.author as book_author, b.cover_url as book_cover_url
-    FROM highlights h
-    LEFT JOIN books b ON h.book_id = b.id
-    ORDER BY h.created_at DESC
-    LIMIT ?`,
-    [limit],
-  );
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    cfi: r.cfi,
-    text: r.text,
-    color: r.color as Highlight["color"],
-    note: r.note || undefined,
-    chapterTitle: r.chapter_title || undefined,
-    createdAt: r.created_at,
-    updatedAt: r.updated_at,
-    bookTitle: r.book_title || "",
-    bookAuthor: r.book_author || "",
-    bookCoverUrl: r.book_cover_url || undefined,
-  }));
-}
-
-/** Get highlight statistics */
-export async function getHighlightStats(): Promise<{
-  totalHighlights: number;
-  highlightsWithNotes: number;
-  totalBooks: number;
-  colorDistribution: Record<string, number>;
-  recentCount: number; // last 7 days
-}> {
-  const database = await getDB();
-
-  const totalRows = await database.select<{ count: number }>(
-    "SELECT COUNT(*) as count FROM highlights",
-  );
-  const notesRows = await database.select<{ count: number }>(
-    "SELECT COUNT(*) as count FROM highlights WHERE note IS NOT NULL AND note != ''",
-  );
-  const booksRows = await database.select<{ count: number }>(
-    "SELECT COUNT(DISTINCT book_id) as count FROM highlights",
-  );
-
-  const colorRows = await database.select<{ color: string; count: number }>(
-    "SELECT color, COUNT(*) as count FROM highlights GROUP BY color",
-  );
-  const colorDistribution: Record<string, number> = {};
-  for (const row of colorRows) {
-    colorDistribution[row.color] = row.count;
-  }
-
-  const sevenDaysAgo = Date.now() - 7 * 24 * 60 * 60 * 1000;
-  const recentRows = await database.select<{ count: number }>(
-    "SELECT COUNT(*) as count FROM highlights WHERE created_at >= ?",
-    [sevenDaysAgo],
-  );
-
-  return {
-    totalHighlights: totalRows[0]?.count || 0,
-    highlightsWithNotes: notesRows[0]?.count || 0,
-    totalBooks: booksRows[0]?.count || 0,
-    colorDistribution,
-    recentCount: recentRows[0]?.count || 0,
-  };
-}
-
-export async function insertHighlight(highlight: Highlight): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "highlights");
-  await database.execute(
-    "INSERT INTO highlights (id, book_id, cfi, text, color, note, chapter_title, created_at, updated_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      highlight.id,
-      highlight.bookId,
-      highlight.cfi,
-      highlight.text,
-      highlight.color,
-      highlight.note || null,
-      highlight.chapterTitle || null,
-      highlight.createdAt,
-      highlight.updatedAt,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function updateHighlight(id: string, updates: Partial<Highlight>): Promise<void> {
-  const database = await getDB();
-  const sets: string[] = [];
-  const values: unknown[] = [];
-
-  if (updates.color !== undefined) {
-    sets.push("color = ?");
-    values.push(updates.color);
-  }
-  if (Object.prototype.hasOwnProperty.call(updates, "note")) {
-    sets.push("note = ?");
-    values.push(updates.note ?? null);
-  }
-  if (updates.text !== undefined) {
-    sets.push("text = ?");
-    values.push(updates.text);
-  }
-  // Add sync tracking
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "highlights");
-  const updatedAt = await nextUpdatedAt(database, "highlights", id);
-  sets.push("updated_at = ?");
-  values.push(updatedAt);
-  sets.push("sync_version = ?");
-  values.push(syncVersion);
-  sets.push("last_modified_by = ?");
-  values.push(deviceId);
-
-  if (sets.length === 0) return;
-  values.push(id);
-  await database.execute(`UPDATE highlights SET ${sets.join(", ")} WHERE id = ?`, values);
-}
-
-export async function deleteHighlight(id: string): Promise<void> {
-  const database = await getDB();
-  await insertTombstone(database, id, "highlights");
-  await database.execute("DELETE FROM highlights WHERE id = ?", [id]);
-}
-
-// --- Notes ---
-
-export async function getNotes(bookId: string): Promise<Note[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    highlight_id: string | null;
-    cfi: string | null;
-    title: string;
-    content: string;
-    chapter_title: string | null;
-    tags: string;
-    created_at: number;
-    updated_at: number;
-  }>("SELECT * FROM notes WHERE book_id = ? ORDER BY created_at DESC", [bookId]);
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    highlightId: r.highlight_id || undefined,
-    cfi: r.cfi || undefined,
-    title: r.title,
-    content: r.content,
-    chapterTitle: r.chapter_title || undefined,
-    tags: parseJSON(r.tags, []),
-    createdAt: r.created_at,
-    updatedAt: r.updated_at,
-  }));
-}
-
-/** Get all notes across all books (for general chat without bookId) */
-export async function getAllNotes(limit = 50): Promise<Note[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    highlight_id: string | null;
-    cfi: string | null;
-    title: string;
-    content: string;
-    chapter_title: string | null;
-    tags: string;
-    created_at: number;
-    updated_at: number;
-  }>("SELECT * FROM notes ORDER BY created_at DESC LIMIT ?", [limit]);
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    highlightId: r.highlight_id || undefined,
-    cfi: r.cfi || undefined,
-    title: r.title,
-    content: r.content,
-    chapterTitle: r.chapter_title || undefined,
-    tags: parseJSON(r.tags, []),
-    createdAt: r.created_at,
-    updatedAt: r.updated_at,
-  }));
-}
-
-export async function insertNote(note: Note): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "notes");
-  await database.execute(
-    "INSERT INTO notes (id, book_id, highlight_id, cfi, title, content, chapter_title, tags, created_at, updated_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      note.id,
-      note.bookId,
-      note.highlightId || null,
-      note.cfi || null,
-      note.title,
-      note.content,
-      note.chapterTitle || null,
-      JSON.stringify(note.tags),
-      note.createdAt,
-      note.updatedAt,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function updateNote(id: string, updates: Partial<Note>): Promise<void> {
-  const database = await getDB();
-  const sets: string[] = [];
-  const values: unknown[] = [];
-
-  if (updates.title !== undefined) {
-    sets.push("title = ?");
-    values.push(updates.title);
-  }
-  if (updates.content !== undefined) {
-    sets.push("content = ?");
-    values.push(updates.content);
-  }
-  if (updates.tags !== undefined) {
-    sets.push("tags = ?");
-    values.push(JSON.stringify(updates.tags));
-  }
-  // Add sync tracking
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "notes");
-  const updatedAt = await nextUpdatedAt(database, "notes", id);
-  sets.push("updated_at = ?");
-  values.push(updatedAt);
-  sets.push("sync_version = ?");
-  values.push(syncVersion);
-  sets.push("last_modified_by = ?");
-  values.push(deviceId);
-
-  if (sets.length === 0) return;
-  values.push(id);
-  await database.execute(`UPDATE notes SET ${sets.join(", ")} WHERE id = ?`, values);
-}
-
-export async function deleteNote(id: string): Promise<void> {
-  const database = await getDB();
-  await insertTombstone(database, id, "notes");
-  await database.execute("DELETE FROM notes WHERE id = ?", [id]);
-}
-
-// --- Bookmarks ---
-
-export async function getBookmarks(bookId: string): Promise<Bookmark[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    cfi: string;
-    label: string | null;
-    chapter_title: string | null;
-    created_at: number;
-  }>("SELECT * FROM bookmarks WHERE book_id = ? ORDER BY created_at DESC", [bookId]);
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    cfi: r.cfi,
-    label: r.label || undefined,
-    chapterTitle: r.chapter_title || undefined,
-    createdAt: r.created_at,
-  }));
-}
-
-export async function insertBookmark(bookmark: Bookmark): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "bookmarks");
-  await database.execute(
-    "INSERT INTO bookmarks (id, book_id, cfi, label, chapter_title, created_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      bookmark.id,
-      bookmark.bookId,
-      bookmark.cfi,
-      bookmark.label || null,
-      bookmark.chapterTitle || null,
-      bookmark.createdAt,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function deleteBookmark(id: string): Promise<void> {
-  const database = await getDB();
-  await insertTombstone(database, id, "bookmarks");
-  await database.execute("DELETE FROM bookmarks WHERE id = ?", [id]);
-}
-
-// --- Threads ---
-
-export async function getThreads(bookId?: string): Promise<Thread[]> {
-  const database = await getDB();
-  const rows = bookId
-    ? await database.select<{
-        id: string;
-        book_id: string | null;
-        title: string;
-        created_at: number;
-        updated_at: number;
-      }>("SELECT * FROM threads WHERE book_id = ? ORDER BY updated_at DESC", [bookId])
-    : await database.select<{
-        id: string;
-        book_id: string | null;
-        title: string;
-        created_at: number;
-        updated_at: number;
-      }>("SELECT * FROM threads ORDER BY updated_at DESC");
-
-  const threads: Thread[] = [];
-  for (const row of rows) {
-    const messages = await getMessages(row.id);
-    threads.push({
-      id: row.id,
-      bookId: row.book_id || undefined,
-      title: row.title,
-      messages,
-      createdAt: row.created_at,
-      updatedAt: row.updated_at,
-    });
-  }
-  return threads;
-}
-
-export async function getThread(id: string): Promise<Thread | null> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string | null;
-    title: string;
-    created_at: number;
-    updated_at: number;
-  }>("SELECT * FROM threads WHERE id = ?", [id]);
-  if (rows.length === 0) return null;
-
-  const row = rows[0];
-  const messages = await getMessages(row.id);
-  return {
-    id: row.id,
-    bookId: row.book_id || undefined,
-    title: row.title,
-    messages,
-    createdAt: row.created_at,
-    updatedAt: row.updated_at,
-  };
-}
-
-export async function insertThread(thread: Thread): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "threads");
-  await database.execute(
-    "INSERT INTO threads (id, book_id, title, created_at, updated_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?)",
-    [
-      thread.id,
-      thread.bookId || null,
-      thread.title,
-      thread.createdAt,
-      thread.updatedAt,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function updateThreadTitle(id: string, title: string): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "threads");
-  const updatedAt = await nextUpdatedAt(database, "threads", id);
-  await database.execute(
-    "UPDATE threads SET title = ?, updated_at = ?, sync_version = ?, last_modified_by = ? WHERE id = ?",
-    [title, updatedAt, syncVersion, deviceId, id],
-  );
-}
-
-export async function deleteThread(id: string): Promise<void> {
-  const database = await getDB();
-  // Get all message IDs in this thread for tombstones
-  const messages = await database.select<{ id: string }>(
-    "SELECT id FROM messages WHERE thread_id = ?",
-    [id],
-  );
-  for (const msg of messages) {
-    await insertTombstone(database, msg.id, "messages");
-  }
-  await insertTombstone(database, id, "threads");
-  await database.execute("DELETE FROM messages WHERE thread_id = ?", [id]);
-  await database.execute("DELETE FROM threads WHERE id = ?", [id]);
-}
-
-export async function deleteThreadsByBookId(bookId: string): Promise<void> {
-  const database = await getDB();
-  // Get all thread IDs for this book
-  const threads = await database.select<{ id: string }>(
-    "SELECT id FROM threads WHERE book_id = ?",
-    [bookId],
-  );
-  // Delete each thread (this handles messages and tombstones)
-  for (const thread of threads) {
-    await deleteThread(thread.id);
-  }
-}
-
-// --- Messages ---
-
-export async function getMessages(threadId: string): Promise<Message[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    thread_id: string;
-    role: string;
-    content: string;
-    citations: string | null;
-    tool_calls: string | null;
-    reasoning: string | null;
-    parts_order: string | null;
-    created_at: number;
-  }>("SELECT * FROM messages WHERE thread_id = ? ORDER BY created_at ASC", [threadId]);
-  return rows.map((r) => ({
-    id: r.id,
-    threadId: r.thread_id,
-    role: r.role as Message["role"],
-    content: r.content,
-    citations: parseJSON(r.citations, undefined),
-    toolCalls: parseJSON(r.tool_calls, undefined),
-    reasoning: parseJSON(r.reasoning, undefined),
-    partsOrder: parseJSON(r.parts_order, undefined),
-    createdAt: r.created_at,
-  }));
-}
-
-export async function insertMessage(message: Message): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "messages");
-  await database.execute(
-    "INSERT INTO messages (id, thread_id, role, content, citations, tool_calls, reasoning, parts_order, created_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      message.id,
-      message.threadId,
-      message.role,
-      message.content,
-      message.citations ? JSON.stringify(message.citations) : null,
-      message.toolCalls ? JSON.stringify(message.toolCalls) : null,
-      message.reasoning ? JSON.stringify(message.reasoning) : null,
-      (message as any).partsOrder ? JSON.stringify((message as any).partsOrder) : null,
-      message.createdAt,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-// --- Reading Sessions ---
-
-export async function getReadingSessions(bookId: string): Promise<ReadingSession[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    started_at: number;
-    ended_at: number | null;
-    total_active_time: number;
-    pages_read: number;
-    state: string;
-  }>("SELECT * FROM reading_sessions WHERE book_id = ? ORDER BY started_at DESC", [bookId]);
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    startedAt: r.started_at,
-    endedAt: r.ended_at || undefined,
-    totalActiveTime: r.total_active_time,
-    pagesRead: r.pages_read,
-    state: r.state as ReadingSession["state"],
-  }));
-}
-
-export async function getReadingSessionsByDateRange(
-  startDate: Date,
-  endDate: Date,
-): Promise<ReadingSession[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    started_at: number;
-    ended_at: number | null;
-    total_active_time: number;
-    pages_read: number;
-    state: string;
-  }>(
-    "SELECT * FROM reading_sessions WHERE started_at >= ? AND started_at <= ? ORDER BY started_at DESC",
-    [startDate.getTime(), endDate.getTime()],
-  );
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    startedAt: r.started_at,
-    endedAt: r.ended_at || undefined,
-    totalActiveTime: r.total_active_time,
-    pagesRead: r.pages_read,
-    state: r.state as ReadingSession["state"],
-  }));
-}
-
-export async function insertReadingSession(session: ReadingSession): Promise<void> {
-  const database = await getDB();
-  const now = Date.now();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "reading_sessions");
-  await database.execute(
-    "INSERT INTO reading_sessions (id, book_id, started_at, ended_at, total_active_time, pages_read, state, updated_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      session.id,
-      session.bookId,
-      session.startedAt,
-      session.endedAt || null,
-      session.totalActiveTime,
-      session.pagesRead,
-      session.state,
-      now,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function updateReadingSession(
-  id: string,
-  updates: Partial<ReadingSession>,
-): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "reading_sessions");
-  const sets: string[] = [];
-  const values: unknown[] = [];
-
-  if (updates.endedAt !== undefined) {
-    sets.push("ended_at = ?");
-    values.push(updates.endedAt);
-  }
-  if (updates.totalActiveTime !== undefined) {
-    sets.push("total_active_time = ?");
-    values.push(updates.totalActiveTime);
-  }
-  if (updates.pagesRead !== undefined) {
-    sets.push("pages_read = ?");
-    values.push(updates.pagesRead);
-  }
-  if (updates.state !== undefined) {
-    sets.push("state = ?");
-    values.push(updates.state);
-  }
-
-  if (sets.length === 0) return;
-
-  const updatedAt = await nextUpdatedAt(database, "reading_sessions", id);
-  sets.push("updated_at = ?");
-  values.push(updatedAt);
-  sets.push("sync_version = ?");
-  values.push(syncVersion);
-  sets.push("last_modified_by = ?");
-  values.push(deviceId);
-
-  values.push(id);
-  await database.execute(`UPDATE reading_sessions SET ${sets.join(", ")} WHERE id = ?`, values);
-}
-
-// --- Chunks ---
-
-export async function getChunks(bookId: string): Promise<Chunk[]> {
-  const database = await getLocalDB();
-  const rows = await database.select<{
-    id: string;
-    book_id: string;
-    chapter_index: number;
-    chapter_title: string;
-    content: string;
-    token_count: number;
-    start_cfi: string | null;
-    end_cfi: string | null;
-    segment_cfis: string | null;
-    embedding: unknown;
-  }>("SELECT * FROM chunks WHERE book_id = ? ORDER BY chapter_index, id", [bookId]);
-  return rows.map((r) => ({
-    id: r.id,
-    bookId: r.book_id,
-    chapterIndex: r.chapter_index,
-    chapterTitle: r.chapter_title,
-    content: r.content,
-    tokenCount: r.token_count,
-    startCfi: r.start_cfi || "",
-    endCfi: r.end_cfi || "",
-    segmentCfis: r.segment_cfis ? JSON.parse(r.segment_cfis) : undefined,
-    embedding: deserializeEmbedding(r.embedding),
-  }));
-}
-
-export async function insertChunks(chunks: Chunk[]): Promise<void> {
-  const database = await getLocalDB();
-  const now = Date.now();
-  for (const chunk of chunks) {
-    await database.execute(
-      "INSERT INTO chunks (id, book_id, chapter_index, chapter_title, content, token_count, start_cfi, end_cfi, segment_cfis, embedding, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-      [
-        chunk.id,
-        chunk.bookId,
-        chunk.chapterIndex,
-        chunk.chapterTitle,
-        chunk.content,
-        chunk.tokenCount,
-        chunk.startCfi || null,
-        chunk.endCfi || null,
-        chunk.segmentCfis ? JSON.stringify(chunk.segmentCfis) : null,
-        serializeEmbedding(chunk.embedding),
-        now,
-      ],
-    );
-  }
-}
-
-export async function deleteChunks(bookId: string): Promise<void> {
-  const database = await getLocalDB();
-  await database.execute("DELETE FROM chunks WHERE book_id = ?", [bookId]);
-}
-
-export async function clearVectorizationFlagsWithoutLocalChunks(): Promise<void> {
-  const database = await getDB();
-  const localDatabase = await getLocalDB();
-  const rows = await localDatabase.select<{ book_id: string }>(
-    "SELECT DISTINCT book_id FROM chunks",
-  );
-  const bookIds = rows.map((row) => row.book_id).filter((bookId) => !!bookId);
-
-  if (bookIds.length === 0) {
-    await database.execute(
-      "UPDATE books SET is_vectorized = 0, vectorize_progress = 0 WHERE is_vectorized != 0 OR vectorize_progress != 0",
-    );
-    return;
-  }
-
-  const batchSize = 400;
-  const clauses: string[] = [];
-  const params: string[] = [];
-
-  for (let index = 0; index < bookIds.length; index += batchSize) {
-    const batch = bookIds.slice(index, index + batchSize);
-    clauses.push(`id NOT IN (${batch.map(() => "?").join(", ")})`);
-    params.push(...batch);
-  }
-
-  await database.execute(
-    `UPDATE books
-     SET is_vectorized = 0, vectorize_progress = 0
-     WHERE (is_vectorized != 0 OR vectorize_progress != 0)
-       AND ${clauses.join(" AND ")}`,
-    params,
-  );
-
-  const restoreClauses: string[] = [];
-  const restoreParams: string[] = [];
-
-  for (let index = 0; index < bookIds.length; index += batchSize) {
-    const batch = bookIds.slice(index, index + batchSize);
-    restoreClauses.push(`id IN (${batch.map(() => "?").join(", ")})`);
-    restoreParams.push(...batch);
-  }
-
-  await database.execute(
-    `UPDATE books
-     SET is_vectorized = 1, vectorize_progress = 1
-     WHERE (is_vectorized = 0 OR vectorize_progress < 1)
-       AND (${restoreClauses.join(" OR ")})`,
-    restoreParams,
-  );
-}
-
-// --- Skills ---
-
-export async function getSkills(): Promise<Skill[]> {
-  const database = await getDB();
-  const rows = await database.select<{
-    id: string;
-    name: string;
-    description: string;
-    icon: string | null;
-    enabled: number;
-    parameters: string;
-    prompt: string;
-    built_in: number;
-    created_at: number;
-    updated_at: number;
-  }>("SELECT * FROM skills ORDER BY created_at ASC");
-  return rows.map((r) => ({
-    id: r.id,
-    name: r.name,
-    description: r.description,
-    icon: r.icon || undefined,
-    enabled: r.enabled === 1,
-    parameters: parseJSON(r.parameters, []),
-    prompt: r.prompt,
-    builtIn: r.built_in === 1,
-    createdAt: r.created_at,
-    updatedAt: r.updated_at,
-  }));
-}
-
-export async function insertSkill(skill: Skill): Promise<void> {
-  const database = await getDB();
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "skills");
-  await database.execute(
-    "INSERT INTO skills (id, name, description, icon, enabled, parameters, prompt, built_in, created_at, updated_at, sync_version, last_modified_by) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      skill.id,
-      skill.name,
-      skill.description,
-      skill.icon || null,
-      skill.enabled ? 1 : 0,
-      JSON.stringify(skill.parameters),
-      skill.prompt,
-      skill.builtIn ? 1 : 0,
-      skill.createdAt,
-      skill.updatedAt,
-      syncVersion,
-      deviceId,
-    ],
-  );
-}
-
-export async function updateSkill(id: string, updates: Partial<Skill>): Promise<void> {
-  const database = await getDB();
-  const sets: string[] = [];
-  const values: unknown[] = [];
-
-  if (updates.name !== undefined) {
-    sets.push("name = ?");
-    values.push(updates.name);
-  }
-  if (updates.description !== undefined) {
-    sets.push("description = ?");
-    values.push(updates.description);
-  }
-  if (updates.enabled !== undefined) {
-    sets.push("enabled = ?");
-    values.push(updates.enabled ? 1 : 0);
-  }
-  if (updates.parameters !== undefined) {
-    sets.push("parameters = ?");
-    values.push(JSON.stringify(updates.parameters));
-  }
-  if (updates.prompt !== undefined) {
-    sets.push("prompt = ?");
-    values.push(updates.prompt);
-  }
-  // Add sync tracking
-  const deviceId = await getDeviceId();
-  const syncVersion = await nextSyncVersion(database, "skills");
-  const updatedAt = await nextUpdatedAt(database, "skills", id);
-  sets.push("updated_at = ?");
-  values.push(updatedAt);
-  sets.push("sync_version = ?");
-  values.push(syncVersion);
-  sets.push("last_modified_by = ?");
-  values.push(deviceId);
-
-  if (sets.length === 0) return;
-  values.push(id);
-  await database.execute(`UPDATE skills SET ${sets.join(", ")} WHERE id = ?`, values);
-}
-
-export async function deleteSkill(id: string): Promise<void> {
-  const database = await getDB();
-  await insertTombstone(database, id, "skills");
-  await database.execute("DELETE FROM skills WHERE id = ?", [id]);
-}
+// --- Core infrastructure ---
+export {
+  getActiveDataRoot,
+  getDatabaseFilePath,
+  cleanupOrphanedSyncRows,
+  getDB,
+  getLocalDB,
+  closeDB,
+  closeLocalDB,
+  ensureNoTransaction,
+  resetDBCache,
+  resetLocalDBCache,
+  getDeviceId,
+  initDatabase,
+  initLocalDatabase,
+  // Shared utilities (used by sync-engine, etc.)
+  nextSyncVersion,
+  nextUpdatedAt,
+  insertTombstone,
+  parseJSON,
+  serializeEmbedding,
+  deserializeEmbedding,
+} from "./db-core";
+
+// --- Domain queries ---
+export {
+  getBooks,
+  getBook,
+  insertBook,
+  updateBook,
+  setBookSyncStatus,
+  deleteBook,
+} from "./book-queries";
+
+export {
+  getHighlights,
+  getAllHighlights,
+  getAllHighlightsWithBooks,
+  getHighlightStats,
+  insertHighlight,
+  updateHighlight,
+  deleteHighlight,
+} from "./highlight-queries";
+export type { HighlightWithBook } from "./highlight-queries";
+
+export {
+  getNotes,
+  getAllNotes,
+  insertNote,
+  updateNote,
+  deleteNote,
+} from "./note-queries";
+
+export {
+  getBookmarks,
+  insertBookmark,
+  deleteBookmark,
+} from "./bookmark-queries";
+
+export {
+  getThreads,
+  getThread,
+  insertThread,
+  updateThreadTitle,
+  deleteThread,
+  deleteThreadsByBookId,
+} from "./thread-queries";
+
+export {
+  getMessages,
+  insertMessage,
+} from "./message-queries";
+
+export {
+  getReadingSessions,
+  getReadingSessionsByDateRange,
+  insertReadingSession,
+  updateReadingSession,
+} from "./session-queries";
+
+export {
+  getChunks,
+  insertChunks,
+  deleteChunks,
+  clearVectorizationFlagsWithoutLocalChunks,
+} from "./chunk-queries";
+
+export {
+  getSkills,
+  insertSkill,
+  updateSkill,
+  deleteSkill,
+} from "./skill-queries";
